Move registration method
