+++ cohort-engine/src/main/java/com/ibm/cohort/engine/translation/BaseCqlTranslationProvider.java
@@ -30,17 +30,6 @@ import com.ibm.cohort.engine.LibraryFormat;
 public abstract class BaseCqlTranslationProvider implements CqlTranslationProvider {
 
 	public static final LibraryFormat DEFAULT_TARGET_FORMAT = LibraryFormat.XML;
-
-	public static void registerModelInfo(File modelInfoFile) {
-		ModelInfo modelInfo = JAXB.unmarshal(modelInfoFile, ModelInfo.class);
-		// Force mapping  to FHIR 4.0.1. Consider supporting different versions in the future.
-		// Possibly add support for auto-loading model info files.
-		modelInfo.setTargetVersion("4.0.1");
-		modelInfo.setTargetUrl("http://hl7.org/fhir");
-		org.hl7.elm.r1.VersionedIdentifier modelId = (new org.hl7.elm.r1.VersionedIdentifier()).withId(modelInfo.getName()).withVersion(modelInfo.getVersion());
-		ModelInfoProvider modelProvider = () -> modelInfo;
-		ModelInfoLoader.registerModelInfoProvider(modelId, modelProvider);
-	}
 	
 	public List<Options> getDefaultOptions() {
 		List<CqlTranslator.Options> defaults = new ArrayList<>(CqlTranslatorOptions.defaultOptions().getOptions());
